+++ src/particles/emission/particle_emission.f90
@@ -204,8 +204,8 @@ IMPLICIT NONE
 ! LOCAL VARIABLES
 INTEGER                          :: i,k,iPart,PositionNbr,iInit,IntSample,particle_count
 INTEGER                          :: NbrOfParticle
-INTEGER(KIND=8)                  :: inserted_Particle_iter,inserted_Particle_time
-INTEGER(KIND=8)                  :: inserted_Particle_diff
+INTEGER(KIND=DP)                 :: inserted_Particle_iter,inserted_Particle_time
+INTEGER(KIND=DP)                 :: inserted_Particle_diff
 REAL                             :: PartIns, RandVal1
 REAL                             :: RiseFactor, RiseTime
 #if USE_MPI
